prototype/snapshot-aggregator/requirements.txt
kafka-python==2.0.2
APScheduler==3.10.4
